PAGOSONLINEPPS-32264 Actualizar modelos para procesamiento de tarjetas débito y agregar género del pagador
